🎨 :: rtt_data 자료구조 자료형 변경
# protocol number에 해당하는 key: int -> unsigned short, tick에 해당하는 value: int -> long long

+++ DummyChattingClient/DummyChattingClientService.h
@@ -23,5 +23,5 @@ private:
     std::vector<std::thread> m_job_threads;
 
     std::mutex m_rtt_lock;
-    std::multimap<int, int> m_rtt_data;
+    std::multimap<unsigned short, long long> m_rtt_data;
 };
